+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog for Nautilai
 Fixed:
 ^^^^^^
 - Acquisition failing to run after changing recording location in Save Options
+- Handling of errors related to nautilai.toml and machine.toml files
 
 
 0.1.44 (2024-06-04)